check if we in powerline-shell dir and do reset

+++ unix/replace-my-vim.sh
@@ -10,9 +10,13 @@ cd $BUNDLE_PATH
 cp ${ORIGINAL_PATH}/patch/powerline-shell.patch $BUNDLE_PATH/powerline-shell/
 cd $BUNDLE_PATH/powerline-shell
 rm -rf segments/idle.py
-git reset --hard
-patch -p1 < powerline-shell.patch
-python install.py
+if [ -d "segments" ]
+then
+	echo "Do reset!!!!!"
+	git reset --hard
+	patch -p1 < powerline-shell.patch
+	python install.py
+fi
 cd ${ORIGINAL_PATH}
 
 cp ./dist/ctags_lang ~/.ctags
